Thin3D: use 16-bit indices. Fixes #12898.

+++ GPU/Common/PresentationCommon.cpp
@@ -18,6 +18,7 @@
 #include <cassert>
 #include <cmath>
 #include <set>
+#include <cstdint>
 #include "base/display.h"
 #include "base/timeutil.h"
 #include "file/vfs.h"
@@ -318,7 +319,7 @@ void PresentationCommon::CreateDeviceObjects() {
 
 	vdata_ = draw_->CreateBuffer(sizeof(Vertex) * 8, BufferUsageFlag::DYNAMIC | BufferUsageFlag::VERTEXDATA);
 	// TODO: Use 4 and a strip?  shorts?
-	idata_ = draw_->CreateBuffer(sizeof(int) * 6, BufferUsageFlag::DYNAMIC | BufferUsageFlag::INDEXDATA);
+	idata_ = draw_->CreateBuffer(sizeof(uint16_t) * 6, BufferUsageFlag::DYNAMIC | BufferUsageFlag::INDEXDATA);
 
 	samplerNearest_ = draw_->CreateSamplerState({ TextureFilter::NEAREST, TextureFilter::NEAREST, TextureFilter::NEAREST, 0.0f, TextureAddressMode::CLAMP_TO_EDGE, TextureAddressMode::CLAMP_TO_EDGE, TextureAddressMode::CLAMP_TO_EDGE });
 	samplerLinear_ = draw_->CreateSamplerState({ TextureFilter::LINEAR, TextureFilter::LINEAR, TextureFilter::LINEAR, 0.0f, TextureAddressMode::CLAMP_TO_EDGE, TextureAddressMode::CLAMP_TO_EDGE, TextureAddressMode::CLAMP_TO_EDGE });
@@ -437,7 +438,8 @@ void PresentationCommon::CopyToOutput(OutputFlags flags, int uvRotation, float u
 	// Make sure Direct3D 11 clears state, since we set shaders outside Draw.
 	draw_->BindPipeline(nullptr);
 
-	int indexes[] = { 0, 1, 2, 0, 2, 3 };
+	// Wouldn't it be enough to do this only when creating the index buffer?
+	uint16_t indexes[] = { 0, 1, 2, 0, 2, 3 };
 	draw_->UpdateBuffer(idata_, (const uint8_t *)indexes, 0, sizeof(indexes), Draw::UPDATE_DISCARD);
 
 	// TODO: If shader objects have been created by now, we might have received errors.
